Factor out batch-rename.rs
So far we relied on the presence of the print-rename script. With an
upcoming change we intend to integrate said script into this repository
as well.
With this change we prepare for doing so by renaming main.rs to
batch-rename.rs, to make room for the introduction of additional
binaries.