Factor out rename() function
The batch-rename script currently invokes print-rename as a separate
binary. However, that is unnecessary overhead and complexity. Rather,
now that both live in the same repository we can replace the spawning of
a process with a mere function call.
With this change we factor out the rename() function that takes care of
renaming a single file.