Wait for all renames to finish
So far, we just spawned rename operations and then forgot about them.
However, a tokio::main does not wait for all spawned tasks to finish
before allowing the program to exit. As a result, the last rename
operation occasionally did not finish.
With this change we make sure to collect all spawned task handles and
await them before exiting.